Bug fix and code review.

+++ OrderManagement.Infrastructure/Migrations/OrderDbContextModelSnapshot.cs
@@ -248,8 +248,8 @@ namespace OrderManagement.Infrastructure.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("TotalPrice")
-                        .HasColumnType("int");
+                    b.Property<double>("TotalPrice")
+                        .HasColumnType("float");
 
                     b.HasKey("Id");
